requirements.txt
# Image Alignment Pipeline Requirements

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image processing
tifffile>=2021.11.0
scikit-image>=0.18.0
Pillow>=8.0.0

# Interactive particle selection (napari)
napari[all]>=0.4.15
magicgui>=0.3.0

# Progress bars and visualization
tqdm>=4.62.0
matplotlib>=3.4.0

# Deep learning for GPU alignment
torch>=1.10.0
torchvision>=0.11.0

# Qt backend for napari (usually installed with napari[all])
# PyQt5>=5.12.0  # Already included in napari[all]
# qtpy>=1.9.0    # Already included in napari[all]
